move spinner prints to utils
