+++ cmd/ef_tests/levm/utils.rs
@@ -2,6 +2,7 @@ use crate::types::EFTest;
 use ethrex_core::{types::Genesis, H256};
 use ethrex_storage::{EngineType, Store};
 use ethrex_vm::{evm_state, EvmState};
+use spinoff::Spinner;
 
 pub fn load_initial_state(test: &EFTest) -> (EvmState, H256) {
     let genesis = Genesis::from(test);
@@ -16,4 +17,20 @@ pub fn load_initial_state(test: &EFTest) -> (EvmState, H256) {
         ),
         genesis.get_block().header.compute_block_hash(),
     )
+}
+
+pub fn spinner_update_text_or_print(spinner: &mut Spinner, text: String, no_spinner: bool) {
+    if no_spinner {
+        println!("{}", text);
+    } else {
+        spinner.update_text(text);
+    }
+}
+
+pub fn spinner_success_or_print(spinner: &mut Spinner, text: String, no_spinner: bool) {
+    if no_spinner {
+        println!("{}", text);
+    } else {
+        spinner.success(&text);
+    }
 }
